+++ tests/testthat/test-tm_t_summary_by.R
@@ -41,14 +41,7 @@ test_that("template_summary_by generates correct expressions", {
           na.rm = FALSE,
           na_level = "<Missing>",
           denom = "N_col",
-          .stats = c("n", "mean_sd", "median", "range", "count_fraction"),
-          .formats = c(
-            n = "xx",
-            mean_sd = "xx.xx (xx.xx)",
-            median = "xx.xx",
-            range = "xx.xx - xx.xx",
-            count_fraction = "xx (xx.%)"
-          )
+          .stats = c("n", "mean_sd", "mean_ci", "median", "median_ci", "quantiles", "range", "count_fraction")
         )
     ),
     table = quote({
@@ -102,7 +95,7 @@ test_that("template_summary_by generates correct expressions when `parallel_vars
           vars = c("AVAL", "CHG"),
           na.rm = FALSE,
           denom = "N_col",
-          .stats = c("n", "mean_sd", "median", "range", "count_fraction")
+          .stats = c("n", "mean_sd", "mean_ci", "median", "median_ci", "quantiles", "range", "count_fraction")
         )
     ),
     table = quote({
@@ -174,6 +167,61 @@ test_that("template_summary_by generates correct expressions when `row_groups` i
           label_pos = "topleft"
         ) %>%
         summarize_row_groups(var = "USUBJID", cfun = cfun_unique)
+    ),
+    table = quote({
+      result <- build_table(lyt = lyt, df = anl, alt_counts_df = adsl)
+      result
+    })
+  )
+  expect_equal(result, expected)
+})
+
+test_that("template_summary_by generates correct expressions for customized numeric statistics", {
+  result <- template_summary_by(
+    parentname = "adsl",
+    dataname = "adlb",
+    arm_var = "ARM",
+    id_var = "USUBJID",
+    sum_vars = c("AVAL"),
+    add_total = TRUE,
+    by_vars = c("AVISIT"),
+    na.rm = FALSE,
+    numeric_stats = c("n"),
+    denominator = "N",
+    drop_arm_levels = TRUE
+  )
+
+  expected <- list(
+    data = quote({
+      anl <- adlb %>%
+        df_explicit_na(
+          omit_columns = setdiff(names(adlb), c("AVISIT", "AVAL")),
+          na_level = "<Missing>"
+        )
+      anl <- anl %>% mutate(ARM = droplevels(ARM))
+      arm_levels <- levels(anl[["ARM"]])
+      adsl <- adsl %>% filter(ARM %in% arm_levels)
+      adsl <- adsl %>% mutate(ARM = droplevels(ARM))
+      adsl <- df_explicit_na(adsl, na_level = "")
+    }),
+    layout_prep = quote(split_fun <- drop_split_levels),
+    layout = quote(
+      lyt <- basic_table() %>%
+        split_cols_by("ARM", split_fun = add_overall_level("All Patients", first = FALSE)) %>%
+        add_colcounts() %>%
+        split_rows_by(
+          "AVISIT",
+          split_label = var_labels(adlb)[["AVISIT"]],
+          split_fun = split_fun,
+          label_pos = "topleft"
+        ) %>%
+        summarize_vars(
+          vars = "AVAL",
+          na.rm = FALSE,
+          na_level = "<Missing>",
+          denom = "N_col",
+          .stats = c("n", "count_fraction")
+        )
     ),
     table = quote({
       result <- build_table(lyt = lyt, df = anl, alt_counts_df = adsl)